Libft: work on Makefile, Header file

+++ Libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: fallan <fallan@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/18 15:23:46 by fallan            #+#    #+#             */
-/*   Updated: 2023/10/18 15:23:46 by fallan           ###   ########.fr       */
+/*   Updated: 2023/10/19 16:52:01 by fallan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "Libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	length;
